Add browser module, update dependencies, and improve argument handling
