+++ crates/volt/Cargo.toml
@@ -20,3 +20,6 @@ strum = { version = "0.26.3", features = ["derive"] }
 tap = "1.0.1"
 unicode-truncate = "2.0.0"
 blerp = { path = "../blerp" }
+human-panic = "2.0.2"
+tracing = "0.1.41"
+tracing-subscriber = "0.3.19"
